Reduce the archive_report_ingest_status tests to just GET
