+++ archive/archive_report_ingest_status/src/docker-compose.yml
@@ -1,9 +1,4 @@
-sns:
-  image: wellcome/fake-sns
-  ports:
-    - "9292:9292"
-
 dynamodb:
-  image: peopleperhour/dynamodb
+  image: "peopleperhour/dynamodb"
   ports:
     - "8000:8000"
